refactor(ChessBoardFactory): 초기화 로직 분리
