+++ src/test/java/chess/domain/chesspiece/ChessPieceTest.java
@@ -1,4 +1,4 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
 import static org.assertj.core.api.Assertions.assertThat;
